meta-builder: initial commit for self hosting
Based on the concept of the build applicance:

https://www.yoctoproject.org/tools-resources/projects/build-appliance

but here we aim for a real machine for deployment, a typical
server machine, like a dual xeon (hence using meta-intel to get
a suitable MACHINE).

What is here now is capable of building itself on the deployed
images of its own build.

End goal is to expand the packages and tweak the system such that
everything you'd expect in a normal server (such as tab completion
on command line, man pages etc.) are all present and functioning
just as one would expect on a server install from one of any of
the bigger commercial distributions.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>