devtools: don't add a SRC_URI to gcc bbappend (build errors)
Due to the shared source in work-shared that gcc uses, one can
introduce spurious build errors in gcc builds which are apparently
due to races in patching without proper dependencies.  An example:

https://bugs.launchpad.net/linaro-oe/+bug/1068735

Use a here-document as a workaround to avoid SRC_URI, and escape the
"$" appropriately within, so they don't get evaluated in the install.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>