util-linux: create an all encompassing package group
Currently util-linux is fragmented into lots of small chunks,
and if one is looking to avoid using busybox wholesale, there
is currently no easy way to say "I want all of it".  Here we
create a packagegroup for all of the util-linux chunks, so
that it need not be duplicated by everyone wanting the same
overall goal of removing busybox.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>