builder: remove all of busybox from main rootfs
Busybox has its place in constrained environments, such as low
space/ram and in initrd/initramfs.  However when used interactively,
it can be frustrating when the user finds incompatibilities or
unsupported features/flags as compared to the full version.

Since this layer is intended for interactive use, we here purge all
of busybox from the main rootfs (but it remains in the initrd).
Unfortunately, this is not as trivial as it sounds.  Note the block
in the commit marked with "begin/end busybox removal".  It was done
by forcibly excluding it, and then dealing with the fallout on a
one by one basis, and examining "bitbake -e" output.

After forced removal succeeded, filling in the gaps of the missing
utilities was required (getty, etc. etc.).  Most of these things
live in util-linux and were packaged almost per-command, but were
not installed.  We create a packagegroup for it, which can eventually
be shared with other users who don't want busybox.

One gap that is not covered by util-linux content is the program
"start-stop-daemon" which can be found in "dpkg", hence why it is
added to the package list.  As a development box, "dpkg" may be
useful for inspecting Debian packages anyway.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>