busybox: restore config functionality for initrd use
Now that we no longer have busybox on the main rootfs at all,
we don't need to reduce the config in order to reduce the number
of busybox symlinks on the main rootfs.

So restore the config to what it was prior to us trimming it
down, so that if someone is stuck in the initrd and trying to do
a rescue from there, they can do things like dhcp, netcat, etc.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>