ifupdown: Create volatile dirs and patch pkg to actually work
We need to drop a file in /etc to ensure that /run/network
exists when volatiles are used.

We also have to fix the package to configure properly on non
debian based machines (dpkg-architecture command missing).

Finally, the rules for the methods for loopback devices had
an inverted test, which causes the loopback device to not
be brought up at boot.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>