local.conf: make generic-x86-64 the default
If people copy the sample config and then deploy i7 binaries on
some lower end hardware, they will get confusing SIGILL faults
and generally be left wondering what went wrong.

So here we dial it back to core2 so that most machines made in the
last 10 years will be ok with the binaries generated.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>