builder: enhance hostname setting ability in deployed image.
Make it so a hostname can be chosen for the deployed machine
image.  We used to use HOSTNAME but that is already in use for
the hostname of the building machine ; so choose TARGETNAME to
make it clear we are talking about the hostname of the image
deployed on the target.

We stick the image build date into the hostname to give some
level of overlap avoidance, and to convey the age of the
original deployment.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>