dom0/1/E: add avahi utils
domains created based on the common package list should be discoverable
by default .. otherwise, we can easily be locked out of domain0 and
be unable to control the system.

By ensuring that avahi-daemon and -utils are in the base package list,
we can configure the containers during install to be located.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>