bolt: Add a bitbake recipe to build bolt
Bolt is a key/value store written entirely in go. It is meant to be
used in projects that do not require a full database server.
This library is needed in order to build consul.

Note: Due to expectations by Go in regards to the structure of our
source directory, this recipe also creates a symlink directory which
links to the src folder itself. This puts our src in a
github.com/<publisher>/" hierarchy, and allows our source to properly
reference dependencies internal to the repo itself.

Signed-off-by: Christopher Braga <chris.braga@windriver.com>