go-checkpoint: Add bitbake recipe to build go-checkpoint
Go-checkpoint is an internal go service created specifically for use
in Hashicorp software. It is used by Hashicorp services to check version
information and broadcast security bulletins to users. By default, this
package is necessary for consul to build, but can be removed from the
dependency chain by setting CHECKPOINT_DISABLE=1. Due to go-checkpoint being
designed to communicate specifically with Hashiicorp software and services,
it has no functionality when separated from these systems.

Signed-off-by: Christopher Braga <chris.braga@windriver.com>