dom0: container control
Add the ability of dom0 to launch peer containers based on containers installed
into the lxc "peer" container group.

The autostart script, and systemd service are based on lxc's systemd and autostart
concepts, but are modified to work with nested containers launched from dom0.

The OverC installer ensures that containers are placed in the proper group,
and have the appropriate directories and fstab for the launch.

The systemd service then starts, and autolaunches any required peer launchers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>