local.conf.sample: set DISTRO
My builds kept on producing images with sysvinit as the init system
despite being told that systemd was the expected default throughout
meta-overc. After poking around I found that the DISTRO in the
local.conf.sample still had the default DISTRO set to 'poky' so
changing this to be the expected 'pod'.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>