etcd: Ignore "no GNU hash in elf binary" message
The etcd package currently gets the warning "no GNU hash in elf binary" when
being packaged. While the issue has been identified as a case of go ignoring
LDFLAGS, attempting to manually feed the LDFLAGS variable to go has been
unsuccessful. In order to avoid this from causing the build to fail, the
recipe has been modified to ignore this error

Signed-off-by: Christopher Braga <chris.braga@windriver.com>