serf: Remove compilation of serf binary from serf bitbake recipe
In newer versions of go.net, restrictions have been set such that it can only
be imported as golang.org/x/net. However, as Serf continues to use an older
version of go.net, it imports it through the path code.google.com/p/go.net.
This conflict means that go.net can not be upreved to commits passed Dec 2014
without breaking our ability to compile the Serf binary. As go.net is currently
required for multiple packages, and the Serf binary is currently not in use,
this patch removes the compilation of said binary, and merely makes the source
available to be imported by other packages.

Signed-off-by: Christopher Braga <chris.braga@windriver.com>