dom0-contctl: lxc net veth remote end create bridge if not exist
If network type is veth and the remote end type is bridge or ovs then
if the supplied l2 or ovs bridge through wr.network.<connection>.remote.link
does not exist then connection is not created.

To make lxc net easier to use,  create bridge if it does not exist.

Signed-off-by: Vu Tran <vu.tran@windriver.com>