dom0-contctl: extend networking support
Currently, lxc container config file allows to setup network between a
container being launched and the container where lxc-start command is
invoked.  This capability is quite limited.

It would be more flexible to be able to connect any container to any other
container or to host.  For example, below is typical setup for networking
communication between containers and between container and host.

* [Host - l2 bridge] <-> [c1]

  A network interface in container c1 connects to a regular l2 bridge
  in Host.

* [Host - ovs] <-> [c1]

  A network interface in container c1 connects to a regular ovs bridge
  in Host.

* [Host - direct] <-> [c1]

  A network interface in container c1 connects to a network
  interface in Host.

* [c2 - L2 bridge] <-> [c1]

  A network interface in container c1 connects to a regular L2 bridge
  in container c2.

* [c2 - direct] <-> [c1]

  A network interface in container c1 connects to a network interface
  in container c2.

Signed-off-by: Vu Tran <vu.tran@windriver.com>