dom0-contctl: add exec-lxc-cmd command
The current command "exec-cmd" allow to run a command within the
container.  However if we want to run a lxc-xxxx utils against
a container, the "exec-cmd" will not be able to do so.  For example
if we want to freeze a running container then run either

    dom0_contctl exec-lxc-cmd -- lxc-freeze -n c2
or
    dom0_contctl exec-lxc-cmd -n c2 -- lxc-freeze -n c2

the "-n c2" in the second form can be specified with different container.
This is useful for nested container.

Also modify the descriptions of "-n" option for some commands.

Signed-off-by: Vu Tran <vu.tran@windriver.com>