dom0-contctl: add launch-group command
launch-autostart command will start all the containers with autostart
enable within a autostart group.  The order of container starting can
be specified in config file.

dom0_contctl uses the following 4 options for autostarting:

* wr.start.auto:   "1" for autostart.  Please note this option is different
                   from "lxc.start.auto" option.
* lxc.start.delay: number of seconds to wait before starting next
                   container.
* lxc.start.order: number indicates the order of this container. The larger
                   the order number the higher the priority to start.
* lxc.group:       what autostart group this container belongs to.  Container
                   can belong to multiple groups.  Each group is separated
                   by commas.

For nested container with autolaunch, the parent must be specified in
config option file with "wr.parent" option.

For example to autolaunch 2 groups peer1 and nested1:

    dom0_contctl launch-group -g peer1 -g nested1

Signed-off-by: Vu Tran <vu.tran@windriver.com>