dom0-contctl: add launch command
Currently we have "launch-peer" and "launch-nested" commands,
further to support relaunch and autolaunch commands for launching
nested container.  We need to be able to specify the parent-child
relationship in container config file.   New option

    wr.parent = <parent container name>

to specify the parent container.  When invoke "launch" command
to start container, this option will be read to determine if
the being launched container is peer or nested.

the "launch-peer" and "launch-nested" commands will ignore this
option and behave as before.

Signed-off-by: Vu Tran <vu.tran@windriver.com>