default packages: Add consul and etcd to the base overc image
Add consul and etcd to the default dom0 rootfs so they may form
the base of a service management and discovery daemon. For this,
purpose a new package group has been created to bundle these
packages.

Note: In order to avoid adding consul and etcd to domE, domE has been
reworked to no longer extend dom0.

Signed-off-by: Christopher Braga <chris.braga@windriver.com>