README.install: Add notes for puppet use with the installer
The installation scripts now support the use of puppet during the
creation of the install media and while installing the target. Here we
are adding notes on how puppet is setup for use. For now we are
leaving the use of puppet as an optional feature, at a later point in
time it will most likely be the default and at that time we can remove
what we are adding here and replace it with more details about how to
work with the puppet configuration to achieve specific configuration
goals.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>