containers: split -setup and runtime packagegroups
We need to have the ability to support running containers (packagegroup-containers),
but not have the system services that autostart containers (packagegroup-container-setup)
in any given image.

So we break out -setup and give it a separate packagegroup.

We also update our images to only add -setup to those that need to autolaunch
containers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>