gvfs: prevent samba and its deps from bleeding in via libgphoto2
It is unclear whether we want/need this extra functionality, but
with the commit:

commit e063a4f3bde0633575aabe9ce25777d970b272f0
Author: Max Krummenacher <max.oss.09@gmail.com>
Date:   Sat Aug 1 18:23:19 2015 +0200

    gvfs: add packageconfig for libgphoto2

we started getting samba and all its dependencies, and at the moment,
they won't even build.  So clobber the PACKAGECONFIG it added for
now and we can revisit this later once samba and friends are in a
better state of affairs.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>