meta-overc: change IMAGE_INSTALL_append to IMAGE_INSTALL +=
"IMAGE_INSTALL +=" will override the definition of
"IMAGE_INSTALL ?=" defined in the following inherited
core-image, and this is what we expected.

Otherwise, "IMAGE_INSTALL_append" will append its value
onto "IMAGE_INSTALL ?=" defined in the following inherited
core-image, which includes "packagegroup-base-extend", and
"packagegroup-base-extend" has a depends chain of
packagegroup-base-extend->
    packagegroup-base-bluetooth->
        bluez4

On the other hand, pod-dom1 image has a dependency on bluez5,
thus will cause a confliction when building pod-dom1 and pod-dome
in a same project environment.

This issue is imported by the commit 6693a483 (default packages: Add
consul and etcd to the base overc image)

Signed-off-by: fupan li <fupan.li@windriver.com>