cgmanager: supply distro at configure time
If you don't supply the distro at configure time,
autoconf will attempt to figure it out by calling
lsb_release on the host or probing for the existence
of certain files on the host.  The actual distro doesn't matter,
but the probing around can cause error messages because
it's a violation of cross compilation principles.

This problem was masked if you happened to have the lsb_release
command on your compile host.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>