etcd: fix error due to '/usr/bin/bash' missing dependency
Building an image which contains etcd can produce the following error:
error: Can't install etcd-git-r0.1@core2_64: no package provides /usr/bin/bash

Looking closely at the etcd sources we find that one of the contained
scripts references '/usr/bin/bash' instead of '/bin/bash' like the
others do. This results in bitbake adding an automatic dependency
which can't be satisfied and the above failure ensues.

There is a more comprehensive upstream fix for this that we will get
when we uprev to the latest etcd but for now we just address the
single script to get around this issue with minimal impact.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>