pod: Introduce pod-cmd-server
The pod-cmd-server is a generic command processor
meant for taking and acting upon commands from
containers.  Currently two commands are
supported, namely reboot and shutdown (poweroff).
For example, this allows for the propagation of reboot
commands from domE to pod-essential via a command
server running on dom0.

The current communication channel used is a named pipe, also
known as a fifo.  This fifo must be visible in the namespace
of both the command server and the container sending the
command.  That is accomplished through the judicious use
of bind mounts already in place between the pods.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>