pod: Introduce container-shutdown-notifier
On a shutdown or a reboot of a container running
this service, a similar shutdown or reboot command
will be sent to the listening pod-cmd-server.  This
allows for a container such as domE to effect a
shutdown or reboot on dom0, which in turn can
propagate the same action to pod-essential to cause
the shutdown or reboot to happen to the entire system.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>