overc-system-agent: add dom0 support
Modify this package's install to install its files to /opt/${PN} Install
container scripts to /etc/overc/container Add python-flask and some
other packages to RDEPS, this is needed for the REST interface.

API change: since we do not support rollback to a specified
snapshot image, specifying of snapshot is now optional.

To support operations to the container dom0, we modify rollback
and activate to permit updates since this container will still be
active. Overc no longer tests to see if a container is active in itself,
this will be handled within the container script.  We also do not support
start/stop of dom0.

Argument handling in the container script updated to be more flexible.
Update README.container to reflect changes to container script arguments.

Signed-off-by: Amy Fong <amy.fong@windriver.com>