overc-system-agent: change container message handling
Previously, over-system-agent's container communication infrastruction
only displays stdout from the container scripts when the container scripts
exits.

Update overc so that both stdout and stderr from the containers scripts
are available.

Also, modify the message handling so that messages are displayed as
they become available instead of waiting til the process is finished.
Some of the processes can take several minutes to complete, with
this change, it provides the user (at least from overc - not from REST)
a sense that something's going on.

Signed-off-by: Amy Fong <amy.fong@windriver.com>