python-smartpm: add builddep command to install all of the build required packages
This command will figure out of the build required packages
for a given srpm package and try to install all of these
packages in the system. If a new version of an already installed
package is available, it will be selected for installation.
BTW, this patch only implemented the cache.py, not touched the ccache.c.

For more information about its usage, please run:
$smart builddep -h.

Signed-off-by: fupan li <fupan.li@windriver.com>
Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>