pod -> cube
'pods' are used in many projects and places. To clear up the namespace, we now
call the building blocks of an OverC system 'cubes'.

This switches the images, variables and READMEs to talk about 'cubes'.

We also move the distro conf to meta-overc, and ensure the distro is "overc".

And finally, the meta-cube layer priority is bumped to ensure its configs take
precence.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>