US68701 : Add support for license acceptance
Within essentials at initial boot time, if there is an executable
script to accept license terms in /etc/accept_license , then run
it before bringing up the system (that is, before pivoting out
of the initramfs environment).

The script can have arbitrary contents, with the caveat that it
must run in the initramfs environment, so it cannot depend on
any functionality not available there.  When the script returns,
the system continues with the normal pivot.

The expected use is for it to check for the presense of a license
file and a license acceptance file.  If the first is present but
the second is absent, then it allows the user to display the
license file or accept the license terms.  Only when the user
has accepted the license terms does it continue.

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>