lxc: remove the dependancy of lxc-net.service from lxc.service
overc essential system used bridge.network to setup the bridge for lxc,
so lxc-net.service is redundant, remove the dependancy from lxc.service
to reduce the boottime.

Another, disable the dmesg output on the console when booting the containers,
and this will make the system's boot console clean and reduce the boottime.

Signed-off-by: fupan li <fupan.li@windriver.com>