dom0-contctl: make the dom0-contctl.service output to serial console
Since we had disabled the dmesg's output after the essential system's
boot up, and dom0-contctl.service will take a longer time to setup
the peer containers such as cube-server on  xilinx-zynq alike arm boards,
there is no messages on the console to notify users what happening at
this time. So change the dom0-contctl.service standard output to serial console
to output some info to tell users the system is not hung.

Signed-off-by: fupan li <fupan.li@windriver.com>