security: add ptys to securetty list
Consoles to other containers are provided via socat, agetty and login.
In order to provide the proper environment for agetty/login, ptys are
created and associated with the login process.

To allow root login, we must add those ptys to secure tty.

To limit resource consumption, and to limit the amount of connections,
we currently only add ~10 ptys. This number can be extended in the
future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>