cube-cmd-server: provide dynamic inter-container consoles
To allow inter-container consoles (and administration) when the
network is not available, we extend the existing command server
infrastructure to create a login service, and named pipes for
communications.

As part of this change, we also change the command server
configuration syntax to yaml, so we can provide new functionality
and options that cannot be easily expressed as tuples.

The syntax of the configuration is as follows:

   command:
      reboot:
        acl:
          - domE
          - cube-server
        action: reboot

      shutdown:
        acl:
          - domE
          - cube-server
        action: shutdown

    monitor:
      container:*

We also extend the command server to listen/monitor multiple
containers. This removes the restrictions of the single server
to client container model, and allows multiple consoles to
be created.

To implement this functionality, the existing command pipes, and
shared container directories are leveraged.

Currently only dom0 is allowed/capable of serving console
conections.

From the end user's perspective, cube-console is invoked as
follows:

 % cube-console dom0

And the login prompt for dom0 appears, using the named pipes
and inter container communication structure.

Behind the scenes the following happens:

 - cube-console writes "console" to the /opt/container/local/cmd-server
   pipe

 - the dom0 cmd-server sees the request, and determins which container
   is requesting a console.

 - If allowed (the acl is checked), socat is used to allocate a pty,
   set permissions, and then invoke agetty to provide login services.

 - A console pipe is created in the requesting container's shared
   directory /opt/container/<container>/dom0-console. The previously
   created agetty is associated with that pipe.

 - cube-console has been waiting, sees the console pipe creation,
   and invokes its own copy of socate in raw mode, to read and
   interact with the console.

 - The user logs in, and the console behaves as expected.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>