cube-console: fix essential console access
The essential layer can be entered in one of two ways: via the
command server console, or via namespaces.

If you enter via the command server, we are completely in the
essential domain, and everything works (including user and
process manipulation).

But we can also more quickly enter via the namespaces (since
/host is available). user and process namespaces cannot be
entered from a child namespace, so they are not available when
we do enter this way .. but the console is useful for working
with the essential rootfs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>