systemd: make udev create or delete device nodes
Since udev 176, the 'devtmpfs' filesystem is required,
and udev will not create or delete device nodes anymore,
it only adjusts permissions and ownership of device nodes
and maintains additional symlinks. But for linux containers,
it usually mount tmpfs to its /dev/, thus the "create" and
"delete" functions are still needed when the containers need
to support devices hot plug.

Signed-off-by: fupan li <fupan.li@windriver.com>