overc: enable peer containers
The OverC system architecture has a container command and control
structure that is consolidated in dom0. The control domain has the
ability to launch VMs and containers.

Containers can be of two flavours: nested or peer.

While launched from dom0, peer containers have their namespaces
directly inherited from the essential layer of the system. This
allows them to operate independently from other containers (and
dom0 itself) from the point of view of resources and constraints.

The initial launch and implementation of container launch only
created peer network namespaces. This is was not the intended design
and was only a stepping stone.

Since the kernel currently does not allow a child namespace to
enter the parent namespaces of proc & user, we must proxy peer
container commands through an agent running within the base/essential
namespace.

The cube-cmd-server provides this agent/proxy, and it is modified
appropriately to execute container management tasks on behalf of
authorized containers (currently only dom0) as specified in the
configuration file.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>