systemd: port udev device creation to 225
The original integration of the systemd udev changes was specific
to 216, and breaks newer versions of udev.

This change ports the patch to 225, and creates version specific
bbappends to work with multiple versions of systemd.

We also change the layer configuration to allow dangling bbappends,
since both versions of systemd will not be present at a given
time.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>