socat: Handle EOF better to prevent IO errors
Socat doesn't currently gracefully handle cases when an EOF is
accompanied by the removal of the file associated with one of its FDs
and a timeout is being used. It will continue to attempt to write data
(if data is supplied) to the FD after the EOF but before the timeout,
which will result in an input/output error. Instead of waiting for the
timeout we can validate the file associated with the FD when we get an
EOF and if the file has been removed we can exit immediately.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>