consoles: allow multiple consoles
The first implementation of the command server only allowed one
connection between two cubes. While not a serious limitation, it
is uncessary.

With this change, a dynamically numbered console pipe is created
and returned to the calling console command. The console then
hooks up to that pipe versus a hard coded name.

We also fork the console in the background on the command server
side, leaving us available to handle more requests.

The combination of the two means that multiple console are now
possible.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>