cube-cmd-server: systemd_228 cgroup fixes
systemd 228 changes the way that cgroups/namespaces are named, which
means that the cube-cmd-server can't detect if it is in a container
or not .. and hence, can't properly start on essential.

We add a new conditional to allow for 228 and 225 to co-exist and
we can once again launch peer containers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>