cube-console: add essential to target list
lxc doesn't know about essential as a valid container, but it is a valid
target for cube-console. To make sure it is visible to users, we test
if we are in dom0, and add essential if we are.

We also add the escape squence "ctrl-\" to essential consoles with this
commit.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>