cube-console: allow nested consoles
Connections between dom0 -> <containers> were using namespaces, or lxc-console.
Either approach had issues. Namespaces do not have a valid environment.
lxc-console can only be exited with a control sequence, and can only have one
active at a time.

To avoid this, we can detect that the target is a container, and execute agetty
to provide a on-demand login shell.

This is not unlike how consoles are created between containers and dom0, but
this technique uses lxc-attach and must always go through dom0, allowing a level
of security and control.

Once established, they are torn down by logging out of the shell, or issuing
the overall cube-console exit sequence (ctrl-\).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>