cube-ctl/console: add explicit error message when invoked from essential
Many of the cube* commands require a command server. When logged into
essential, there is no monitoring server (since we are the root of the system)
and hence the commands don't work.

Commands should either be directly invoked (ie. lxc*) or better, the
control domain used (dom0).

To make this obvious/explicit, we can detect that we are not in a container
and dump an explanation.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>