meta-overc: fix the bug of missing /proc when do upgrading
For do host upgrading, some packages's postinstall need /proc
to do someting, and the previous chroot didn't mount it. This patch
add the mounting of it to chroot rootfs. Another, /var/volatile/tmp
has already existed in the rootfs and there is no need to create it,
thus remove the creating steps.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>