overc-conftools: Add new package to handle runtime system configuration
The 'overc-conftools' package can be used to handle system
configuration tasks that are not viable to complete during image or
installer creation. It is intended to run during boot of -essential,
prior to the launching of any containers. This initial version handles
configuration of the network fabric used by OverC. Configurations can
be applied only on the first boot but also on subsequent boots, it is
up the specific configuration to handle its life-cycle. System level
configuration should exist as additions and modifications to this
package but eventually overc-conftools will allow individual recipes
to contribute their own configurations to be executed by the
overc-conftools system.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>