lxc: use libexecdir instead of libdir
Commit f35b2e29d9bfa817e64b83df11ebbbfe0f4e8e5c [bitbake: set default
libexecdir to $prefix/libexec] in poky changes things such that libs
and executables are separated. This results in 'lxc-containers' not
being found where it was previously and as such the attempt to 'sed'
this file fails with:

.../image//usr/lib64/lxc/lxc/lxc-containers: No such file or directory

Swapping in libexecdir in place of libdir and trimming the path some
lets the sed command succeed. For now this fixes the build issue. If
the runtime has issues with this new split we may have to overwrite
the default libexecdir. For now we will run with it.

To not break builds where commit f35b2e29d9 is yet to be applied we
continue to support the old path.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>