ntp: avoid QA error by removing empty libexecdir
We are currently getting a QA error:

ERROR: QA Issue: ntp: Files/directories were installed but not shipped in any package:
  /usr/libexec

This is a result of the commit f35b2e29d9bf [bitbake: set default
libexecdir to $prefix/libexec] on poky which splits executables and
libraries. Since we configure ntp in such a way that nothing is put in
libexecdir we can avoid the QA error by simply removing the empty
directory.

This fix has been sent to meta-openembedded but we will carry the fix
here until it or another fix is merged to meta-openembedded.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>