lxc: add openvswitch networking scripts
LXC doesn't ship with capabilities to support openvswitch
connectivity. With these scripts we can define an LXC container which
will create (if needed) and connect to an OVS bridge.

To make use of these scripts you need to edit the container
configuration. Comment out the 'lxc.network.link' and add lines for
'lxc.network.script.up' and 'lxc.network.script.down'. For example:

   #lxc.network.link = br0
   lxc.network.name = eth0
   lxc.network.script.up = /etc/lxc/ovs-up
   lxc.network.script.down = /etc/lxc/ovs-down

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>