builder-base.bbclass: add support for OVS networking
Currently the images are modified during rootfs creation to make use
of kernel bridging facilities to allow containers and the -essential
OS to create the network fabric of an OverC based system. Although
this configuration results in a working network topology it has
several downsides. For example it is a fairly rigid topology which
doesn't allow for complexities we anticipate the OverC framework will
require out-of-the-box or by users setting up their own containers
within the OverC framework. It is also susceptible to having the
configuration overwritten during package uprevs during the normal
life-cycle of a device deployment.

Using the init system type along with a newly added
'overc-bridge-networking' DISTRO_FEATURE determine how to setup the
base networking configuration in the rootfs.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>

fixup 82a882df6719f24049c37a3f57c99dcaa4766ce0

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>