dnsmasq: start after systemd-resolved
Attempting to start dnsmasq prior to systemd-resolved will cause the
service to fail to start as dnsmasq requires the presence of the
/etc/resolv.conf file. This file is a volatile and only provided once
systemd-resolved is started.

This issue should also affect meta-openembedded. A patch will be sent
to fix this issue there as it is where the .service file is defined.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>