meta-overc: add a "force" option to rollback the container
The previous rollback function will do nothing with dom0 which means
it will not support dom0's rollback.

In addition to the container's name check, it also check is the container
running, if yes, it'll also do nothing with it.

For this "force", it will skip all of those check, including name check and
running status check switch the container to the top of its snapshot history
list regardless of the status of the container directly.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>