overc-conftools: fixup OVS networking
Remove use of systemd-networkd DHCP opting instead to use dnsmasq,
this allows us to overwrite the gateway to make sure dhcp clients are
using the network prime as the gateway, which is required to enable
access to containers on br-int to get external access.

Allow containers to have static IPs, which can be easily configured as
a network offset in the site.pp file.

Enforce puppet orderring to ensure we restart the network services on
-essential only after we have modified the configurations.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>